Safari-611.1.10/LayoutTests/editing/inserting/smart-quote-with-all-configurations-expected.txt


config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":true,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":true,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":true,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":true,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":false,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":false,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":false,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":true,"smartDash":false,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":true,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":true,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":true,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":true,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":false,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":false,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":false,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":true,"smartLink":false,"smartDash":false,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":true,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":true,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":true,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":true,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":false,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":false,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":false,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":true,"smartDash":false,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":true,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":true,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":true,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":true,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":false,"textReplacement":true,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":false,"textReplacement":true,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":false,"textReplacement":false,"autocorrection":true})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"

config({"smartQuote":true,"continuousSpellchecking":false,"smartLink":false,"smartDash":false,"textReplacement":false,"autocorrection":false})
PASS type("We're"); type(" "); type("good"); editor.textContent is "We’re good"
We’re good
